Move send_ping_pong_messages to remote module
