+++ deps/wazuh_testing/wazuh_testing/remote.py
@@ -100,3 +100,37 @@ def check_syslog_event(wazuh_archives_log_monitor, message, port, protocol):
     detect_archives_log_event(archives_monitor=wazuh_archives_log_monitor,
                               callback=callback_detect_syslog_event(message),
                               error_message="Syslog message wasn't received or took too much time.")
+
+
+def send_ping_pong_messages(protocol, manager_address, port):
+    """This function sends the ping message to the manager
+
+    This message is the first of many between the manager and the agents. It is used to check if both of them are ready
+    to send and receive other messages
+
+    Args:
+        protocol (str): it can be UDP or TCP
+        manager_address (str): address of the manager. IP and hostname are valid options
+        port (int): port where the manager has bound the remoted port
+
+    Returns:
+        bytes: returns the #pong message from the manager
+
+    Raises:
+        ConnectionRefusedError: if there's a problem while sending messages to the manager
+    """
+    if protocol == "UDP":
+        sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+        ping_msg = b'#ping'
+    else:
+        sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+        msg = '#ping'
+        msg_size = len(bytearray(msg, 'utf-8'))
+        # Since the message size's is represented as an uint32, you need to use 4 bytes to represent it
+        ping_msg = msg_size.to_bytes(4, 'little') + msg.encode()
+
+    sock.connect((manager_address, port))
+    sock.send(ping_msg)
+    response = sock.recv(len(ping_msg))
+    sock.close()
+    return response if protocol == "UDP" else response[-5:]
